change rvize sim to universal
